+++ R/data_xform_sample.R
@@ -1,0 +1,16 @@
+#' dataset to demonstrate xform functions
+#'
+#' There is no particular significance to the data
+#'
+#' @format A \code{"dplyr::tbl_df()"} with 10 rows and 5 variables:
+#' \describe{
+#'   \item{dtm_chicago}{POSIXct, timezone set to "America/Chicago"}
+#'   \item{dbl_a}{double, sequence of doubles}
+#'   \item{int_a}{integer, sequence of integers}
+#'   \item{char_a}{character, sequence of character strings}
+#'   \item{dtm_paris}{POSIXct, timezone set to "Europe/Paris"}
+#'   \item{dbl_b}{double, sequence of doubles}
+#'   \item{int_b}{integer, sequence of integers}
+#'   \item{char_b}{character, sequence of character strings}
+#' }
+"xform_sample"