+++ NEWS.md
@@ -1,3 +1,6 @@
 # xformplus 0.0.0.9000
 
-* First collection of functions.
+* adds `df_set_tz()`, sets the timezone for all columns of a dataframe
+* adds `colname_type()`, returns column-names of a dataframe that have a given type.
+
+